Patch by Jan Klötzke (with additional TODO comments): * Add a "fault_callback" to the thread structure which is called when a   unhandled page fault happens in user space. A SIGSEGV will only be sent   if the callback returns "true".
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@25609 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/system/kernel/vm/vm.cpp
@@ -3864,8 +3864,13 @@ vm_page_fault(addr_t address, addr_t faultAddress, bool isWrite, bool isUser,
 			release_sem_etc(addressSpace->sem, READ_COUNT, 0);
 			vm_put_address_space(addressSpace);
 #endif
-			if (user_debug_exception_occurred(B_SEGMENT_VIOLATION, SIGSEGV))
-				send_signal(team_get_current_team_id(), SIGSEGV);
+			struct thread *thread = thread_get_current_thread();
+			// TODO: the fault_callback is a temporary solution for vm86
+			if (thread->fault_callback == NULL
+				|| thread->fault_callback(address, faultAddress, isWrite)) {
+				if (user_debug_exception_occurred(B_SEGMENT_VIOLATION, SIGSEGV))
+					send_signal(team_get_current_team_id(), SIGSEGV);
+			}
 		}
 	}
